Fix: 클라이밍 채널 참여 취소 에러 처리

+++ core/src/main/java/org/bookwoori/core/domain/climbing/facade/ClimbingFacade.java
@@ -60,6 +60,9 @@ public class ClimbingFacade {
         Climbing climbing = climbingService.getClimbingById(climbingId);
         boolean isJoined = climbingMemberService.isJoined(currentMember, climbing);
         if (isJoined) {
+            if (climbingMemberService.isOwner(currentMember, climbing)) {
+                throw new CustomException(ErrorCode.OWNER_CANNOT_LEAVE);
+            }
             climbingMemberService.removeMember(currentMember, climbing);
             return false;
         } else {
